canny chyba działa
